requirements.txt
pygatt==4.0.5
PyYAML==6.0
